feat: Enhance AddOrUpdateText method in AnalysisLinkRepositoryMongo for better document handling and implement Excel export functionality in SearchResultsComponent
