+++ frontend-angular/package.json
@@ -26,7 +26,6 @@
     "datatables.net": "^2.2.2",
     "datatables.net-dt": "^2.2.2",
     "express": "^4.18.2",
-    "file-saver": "^2.0.5",
     "markdown-it": "^14.1.0",
     "postcss": "^8.5.3",
     "rxjs": "~7.8.0",
@@ -40,6 +39,7 @@
     "@angular/cli": "^19.2.8",
     "@angular/compiler-cli": "^19.1.0",
     "@types/express": "^4.17.17",
+    "@types/file-saver": "^2.0.7",
     "@types/jasmine": "~5.1.0",
     "@types/markdown-it": "^14.1.2",
     "@types/node": "^18.18.0",
