+++ frontend-angular/src/app/components/search-results/search-results.component.ts
@@ -2,6 +2,9 @@ import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ModalAnalysisMediaSocialComponent } from '../modal-analysis-media-social/modal-analysis-media-social.component';
 
+import * as XLSX from 'xlsx';
+import * as FileSaver from 'file-saver';
+
 @Component({
   selector: 'app-search-results',
   standalone: true,
@@ -31,4 +34,31 @@ export class SearchResultsComponent {
     const supportServices: string[] = ['tiktok.com', 'facebook.com'];
     return supportServices.some((ss) => link.includes(ss));
   }
+  // #region [Xử lí xuất dữ liệu tìm kiếm ra file excel]
+  exportToExcel(): void {
+    // Chọn dữ liệu muốn xuất, ví dụ: this.searchResultsList hoặc this.keywordModels
+    const data = this.searchResultsList.map((item) => ({
+      'Từ khóa': item.siteSearch,
+      'Kết quả': item.showText,
+      // Thêm các trường khác nếu muốn
+    }));
+
+    // Tạo worksheet và workbook
+    const worksheet: XLSX.WorkSheet = XLSX.utils.json_to_sheet(data);
+    const workbook: XLSX.WorkBook = {
+      Sheets: { 'Kết quả tìm kiếm': worksheet },
+      SheetNames: ['Kết quả tìm kiếm'],
+    };
+
+    // Xuất file
+    const excelBuffer: any = XLSX.write(workbook, {
+      bookType: 'xlsx',
+      type: 'array',
+    });
+    const blob: Blob = new Blob([excelBuffer], {
+      type: 'application/octet-stream',
+    });
+    FileSaver.saveAs(blob, 'ket-qua-tim-kiem.xlsx');
+  }
+  // #endregion
 }
